Fix some deprecation warnings
Fixed
* conftest.py:30: DeprecationWarning: the imp module is deprecated in favour of importlib; see the module's documentation for alternative uses
* mslib/msui/mscolab_project.py:486: DeprecationWarning: an integer is required (got type float).  Implicit conversion to integers using __int__ is deprecated, and may be removed in a future version of Python.
* mslib/msui/mscolab_project.py:647: DeprecationWarning: an integer is required (got type float).  Implicit conversion to integers using __int__ is deprecated, and may be removed in a future version of Python.
* mslib/msui/topview.py:354: DeprecationWarning: In future, it will be an error for 'np.bool_' scalars to be interpreted as an index

See issue #863

+++ mslib/msui/mscolab_project.py
@@ -484,7 +484,7 @@ class MessageItem(QtWidgets.QWidget):
         text_browser.anchorClicked.connect(self.on_link_click)
         text_browser.show()
         text_browser.setFixedHeight(
-            text_browser.document().size().height() + text_browser.contentsMargins().top() * 2
+            int(text_browser.document().size().height() + text_browser.contentsMargins().top() * 2)
         )
         return text_browser
 
@@ -645,7 +645,7 @@ class MessageItem(QtWidgets.QWidget):
         html = self.chat_window.markdown.convert(self.message_text)
         self.messageBox.setHtml(html)
         self.messageBox.setFixedHeight(
-            self.messageBox.document().size().height() + self.messageBox.contentsMargins().top() * 2
+            int(self.messageBox.document().size().height() + self.messageBox.contentsMargins().top() * 2)
         )
         self.textArea.adjustSize()
